app_table/requirements.txt
asgiref==3.6.0
blinker==1.6.2
click==8.1.3
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pytz==2023.3
sqlparse==0.4.4
Werkzeug==2.3.3
